tests/wpt/tests/css/css-borders/corner-shape/corner-shape-outside-left.html

<!DOCTYPE html>
<title>shape-outside with corner-shape: notch</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#shapes-from-box-values">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shape">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
body {
  margin: 0;
}
.container {
  width: 200px;
  line-height: 0;
}

.shape {
  float: left;
  shape-outside: border-box;
  border-radius: 48%;
  box-sizing: content-box;
  height: 40px;
  width: 40px;
  padding: 20px;
  border: 20px solid lightgreen;
  margin: 20px;
  background-color: orange;
}

.box {
  display: inline-block;
  width: 60px;
  background-color: blue;
}

.longbox {
  display: inline-block;
  width: 200px;
  height: 20px;
  background-color: blue;
}
</style>

<main class="container">
  <div class="shape"></div>
  <div class="longbox"></div> <!-- Saturate the margin space -->
  <div class="box" style="height: 24px;"></div> <!-- Box at corner -->
  <div class="box" style="height: 36px;"></div>
  <div class="box" style="height: 36px;"></div>
  <div class="box" style="height: 24px;"></div> <!-- Box at corner -->
  <div class="longbox"></div> <!-- Saturate the margin space -->
</main>

<script>
function shape_outside_corner_shape_test(corner_shape, expected) {
  test(() => {
    const shape = document.querySelector(".shape");
    shape.style.setProperty("corner-shape", corner_shape);
    const actual = Array.from(document.querySelectorAll(".container .box")).map(b => b.getBoundingClientRect().x);
    assert_array_approx_equals(actual, expected, 2);
  }, `corner-shape: ${corner_shape} with shape-outside`);
}

shape_outside_corner_shape_test("notch", [82, 140, 140, 82]);
shape_outside_corner_shape_test("bevel", [106, 140, 140, 106]);
shape_outside_corner_shape_test("notch bevel", [106, 140, 140, 82]);
shape_outside_corner_shape_test("round", [130, 140, 140, 130]);
shape_outside_corner_shape_test("square", [140, 140, 140, 140]);
shape_outside_corner_shape_test("scoop", [88, 140, 140, 88]);
shape_outside_corner_shape_test("superellipse(1.5)", [135, 140, 140, 135]);
shape_outside_corner_shape_test("superellipse(-.8)", [88, 140, 140, 88]);
</script>
