LayoutTests/imported/w3c/web-platform-tests/css/css-borders/corner-shape/corner-shape-outside-left-expected.txt


FAIL corner-shape: notch with shape-outside assert_array_approx_equals: property 0, expected 82 +/- 2, expected 82 but got 129.1875
FAIL corner-shape: bevel with shape-outside assert_array_approx_equals: property 0, expected 106 +/- 2, expected 106 but got 129.1875
FAIL corner-shape: notch bevel with shape-outside assert_array_approx_equals: property 0, expected 106 +/- 2, expected 106 but got 129.1875
PASS corner-shape: round with shape-outside
FAIL corner-shape: square with shape-outside assert_array_approx_equals: property 0, expected 140 +/- 2, expected 140 but got 129.1875
FAIL corner-shape: scoop with shape-outside assert_array_approx_equals: property 0, expected 88 +/- 2, expected 88 but got 129.1875
FAIL corner-shape: superellipse(1.5) with shape-outside assert_array_approx_equals: property 0, expected 135 +/- 2, expected 135 but got 129.1875
FAIL corner-shape: superellipse(-.8) with shape-outside assert_array_approx_equals: property 0, expected 88 +/- 2, expected 88 but got 129.1875
